outsourcing collections and groups into an own component
