+++ src/app/components/profile/collections/collections.component.html
@@ -1,0 +1,111 @@
+<!-- COLLECTIONS -->
+<mat-expansion-panel>
+    <mat-expansion-panel-header>
+      <mat-panel-title> {{ 'Collections' | translate }} </mat-panel-title>
+      <mat-panel-description> {{ 'What you’ve worked on' | translate }} </mat-panel-description>
+    </mat-expansion-panel-header>
+    <ng-template matExpansionPanelContent>
+      <div class="tab-main">
+        @if (!showPartakingCompilations) {
+          @if (userCompilations.length === 0) {
+            <div>
+              <p>{{ 'You do not have any collections' | translate }}</p>
+            </div>
+          }
+          <div class="entity-grid" id="main">
+            @for (element of userCompilations; track element) {
+              <div class="grid-item">
+                <app-grid-element
+                  [element]="element"
+                  [disableTypeInfo]="false"
+                  [disableNavigationOnClick]="true"
+                  >
+                </app-grid-element>
+                <!-- Advanced settings -->
+                <button
+                  class="actionbutton"
+                  mat-icon-button
+                  [matMenuTriggerFor]="menu"
+                  color="primary"
+                  >
+                  <mat-icon>more_horiz</mat-icon>
+                </button>
+                <mat-menu #menu="matMenu" yPosition="below">
+                  <button
+                    mat-menu-item
+                    [routerLink]="['/compilation', element._id]"
+                    >
+                    <mat-icon color="primary">remove_red_eye</mat-icon>
+                    {{ 'Visit collection detail page' | translate }}
+                  </button>
+                  <button
+                    mat-menu-item
+                    (click)="openCompilationCreation(element)"
+                    >
+                    <mat-icon color="primary">edit</mat-icon>
+                    {{ 'Edit collection' | translate }}
+                  </button>
+                  <button
+                    mat-menu-item
+                    (click)="removeCompilationDialog(element)"
+                    >
+                    <mat-icon color="warn">delete</mat-icon>
+                    {{ 'Delete collection' | translate }}
+                  </button>
+                </mat-menu>
+              </div>
+            }
+          </div>
+        }
+        @if (showPartakingCompilations) {
+          @if (partakingCompilations.length === 0) {
+            <div>
+              <p>{{ 'You are not partaking in any collections' | translate }}</p>
+            </div>
+          }
+          <div class="entity-grid">
+            @for (element of partakingCompilations; track element) {
+              <app-grid-element
+                class="grid-item"
+                [element]="element"
+                [disableTypeInfo]="false"
+                [disableNavigationOnClick]="true"
+                >
+              </app-grid-element>
+            }
+          </div>
+        }
+      </div>
+      <mat-action-row class="action-row-left-align">
+        <div class="tab-help">
+          <button
+            mat-raised-button
+            color="primary"
+            (click)="openCompilationCreation()"
+            >
+            {{ 'Create a new collection' | translate }}
+          </button>
+          <mat-divider [vertical]="true"></mat-divider>
+          <div class="tab-help-slider-container">
+            <mat-chip-listbox>
+              <mat-chip-option
+                color="accent"
+                (click)="showPartakingCompilations = false"
+                [selected]="!showPartakingCompilations"
+                >{{ 'My collections' | translate }}</mat-chip-option
+                >
+                <mat-slide-toggle
+                  [(ngModel)]="showPartakingCompilations"
+                ></mat-slide-toggle>
+                <mat-chip-option
+                  color="accent"
+                  (click)="showPartakingCompilations = true"
+                  [selected]="showPartakingCompilations"
+                  >{{ 'Collections I partake in' | translate }}</mat-chip-option
+                  >
+                </mat-chip-listbox>
+              </div>
+            </div>
+          </mat-action-row>
+    </ng-template>
+</mat-expansion-panel>
