+++ src/app/components/profile/collections/collections.component.scss
@@ -1,0 +1,128 @@
+@use '@angular/material' as mat;
+
+.profile-page-grid {
+  display: grid;
+  grid-gap: 1rem;
+  grid-template-columns: repeat(2, 1fr);
+  @media (min-width: 1024px) {
+    grid-template-columns: repeat(3, 1fr);
+  }
+  @media (min-width: 1280px) {
+    grid-template-columns: repeat(4, 1fr);
+  }
+  @media (min-width: 1600px) {
+    grid-template-columns: repeat(6, 1fr);
+  }
+
+  .mat-mdc-card {
+    @include mat.elevation-transition;
+    @include mat.elevation(2);
+
+    border-radius: 0;
+
+    &:hover {
+      @include mat.elevation(8);
+    }
+
+    .mat-mdc-card-title {
+      font-weight: lighter;
+      font-size: medium;
+    }
+  }
+}
+
+.grid-item {
+  position: relative;
+  cursor: context-menu;
+
+  .actionbutton {
+    @include mat.elevation(1);
+  }
+}
+
+.group-grid-item {
+  @include mat.elevation(2);
+
+  h1,
+  p {
+    text-align: center;
+  }
+}
+
+.tab-help {
+  * {
+    padding: 0 0.5rem;
+  }
+}
+
+.detail-block {
+  padding: 10px 0;
+}
+
+.actionbutton {
+  position: absolute;
+  top: 16px;
+  left: 16px;
+  z-index: 1;
+  background-color: rgba(255, 255, 255, 0.9);
+  display: grid;
+  place-items: center;
+  width: 24px;
+  height: 24px;
+  line-height: 24px;
+  padding: 0;
+  .material-icons {
+    font-size: 16px;
+    width: 16px;
+    height: 16px;
+    line-height: 16px;
+  }
+}
+
+#profile-page-content {
+  height: 100%;
+  padding: 1rem 2rem 0 2rem;
+  overflow: hidden;
+}
+
+span.mat-content {
+  display: block !important;
+}
+
+.action-row-left-align {
+  justify-content: flex-start;
+}
+
+.content {
+  display: flex;
+  flex-direction: column;
+}
+
+.tab-help {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.mat-mdc-card-actions {
+  display: flex;
+  justify-content: flex-end;
+  align-items: center;
+}
+
+.mat-icon {
+  font-size: 24px;
+}
+
+.mat-mdc-menu-content,
+.mat-mdc-menu-content .mat-mdc-menu-item .mat-mdc-menu-item-text {
+  font-family: 'Open Sans';
+}
+
+.mat-menu-item {
+  font-family: 'Open Sans';
+}
+
+.mat-mdc-menu-item span {
+  font-family: var(--font-stack);
+}
